% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RDataGovAU-package.R
\docType{package}
\name{RDataGovAU}
\alias{RDataGovAU}
\alias{RDataGovAU-package}
\title{RDataGovAU: Obtain Open Datasets From data.gov.au}
\description{
Uses the ckan API on data.gov.au to access, query, and
retrieve open data sets.
}
\section{internal functions}{

The internal yet exported functions ...
}

\section{query functions}{

The query functions ...
}

\section{processing functions}{

The processing functions ...
}

